Refactors virtual spine synthesis into spine.rs data structures
Addresses review feedback: folds virtual_main module into reticulate/spine.rs
as VirtualSpine + Vertebra + SpineLayout with proper methods (build, source,
check_output_collisions). Moves path helpers (to_forward_slash,
sanitize_handle_segment, escape_typst_content) to path_utils.rs. Deduplicates
RheoWorld constructors via init_resources. Renames new_with_virtual_main to
new_for_bundle. Documents why #figure is used for cross-vertebra handle anchors
(#metadata and bare document labels are not referenceable in Typst 0.15).
