+++ crates/core/src/reticulate/mod.rs
@@ -4,3 +4,5 @@ pub mod spine;
 pub mod transformer;
 pub mod types;
 pub mod validator;
+
+pub use spine::{SpineLayout, Vertebra, VirtualSpine};
